requirements.txt
django==1.5.2
django-annoying==0.7.7
django-debug-toolbar
django-nose
django-widget-tweaks==1.3
git+https://github.com/LeadSift/django-gravatar.git
MySQL-python==1.2.4
South==0.8
